Adding @JsonView support for JAX-RS resources.
