+++ src/test/groovy/com/benjaminsproule/swagger/gradleplugin/test/model/TestJsonViewTwo.groovy
@@ -1,0 +1,5 @@
+package com.benjaminsproule.swagger.gradleplugin.test.model
+
+interface TestJsonViewTwo {
+
+}
