--- src/test/groovy/com/benjaminsproule/swagger/gradleplugin/test/springmvc/model/TestJsonViewOne.groovy
@@ -1,5 +1,0 @@
-package com.benjaminsproule.swagger.gradleplugin.test.springmvc.model
-
-interface TestJsonViewOne {
-
-}
