--- src/test/groovy/com/benjaminsproule/swagger/gradleplugin/test/springmvc/model/TestJsonViewTwo.groovy
@@ -1,5 +1,0 @@
-package com.benjaminsproule.swagger.gradleplugin.test.springmvc.model
-
-interface TestJsonViewTwo {
-
-}
